+++ src/components/AddCartButton/AddCartButton.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
+import { AddCartIcon, DeleteCartIcon } from '../../assets';
 import * as S from './AddCartButton.styled';
-import { AddCartIcon, DeleteCartIcon } from '../assets';
 
 interface AddCartButtonProps {
   onAddClick: (event: React.MouseEvent) => void;
